# api/docs/openapi_rapidapi_hub.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api-Desk Whales",
    "version": "2.0.0",
    "description": "Multi-chain whale transfer API (threshold ~> USD): live feed, entity labels, risk score and impact estimate. BASIC/PRO: core whale+stats. ULTRA/MEGA: + timeseries and market analysis. Not a generic price/honeypot aggregator."
  },
  "paths": {
...
    "/v1/market/assets": {
      "get": {
        "tags": [
          "market-signals"
        ],
        "summary": "Market assets",
        "operationId": "marketAssets",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/analysis": {
      "get": {
        "tags": [
          "market-signals"
        ],
        "summary": "Market analysis",
        "operationId": "marketAnalysis",
        "parameters": [
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "BTC",
              "title": "Asset"
            }
          },
          {
            "name": "persist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Persist"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketAnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketAnalysisOut": {
        "properties": {
          "asset": {
            "type": "string",
            "title": "Asset"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "market_microstructure"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "mid_price": {
            "type": "number",
            "title": "Mid Price"
          },
          "stress_score": {
            "type": "number",
            "title": "Stress Score"
          },
          "spillover_hint": {
            "type": "number",
            "title": "Spillover Hint"
          },
          "regime": {
            "type": "string",
            "title": "Regime"
          },
          "components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Components"
          },
          "book_summary": {
            "type": "object",
            "title": "Book Summary"
          },
          "volume_summary": {
            "type": "object",
            "title": "Volume Summary"
          },
          "trades_summary": {
            "type": "object",
            "title": "Trades Summary"
          },
          "venues_summary": {
            "type": "object",
            "title": "Venues Summary"
          },
          "cross_exchange": {
            "type": "object",
            "title": "Cross Exchange"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/MarketSignalOut"
            },
            "type": "array",
            "title": "Signals"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "Se\u00f1ales auxiliares para an\u00e1lisis. No sustituyen detecci\u00f3n de whales."
          }
        },
        "type": "object",
        "required": [
          "asset",
          "source",
          "mid_price",
          "stress_score",
          "spillover_hint",
          "regime",
          "components",
          "book_summary",
          "volume_summary",
          "trades_summary",
          "venues_summary",
          "cross_exchange",
          "signals",
          "fetched_at"
        ],
        "title": "MarketAnalysisOut"
      },
...
      "MarketSignalOut": {
        "properties": {
          "signal_type": {
            "type": "string",
            "title": "Signal Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "global"
          },
          "detail": {
            "type": "object",
            "title": "Detail"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "signal_type",
          "severity",
          "score",
          "title",
          "ts"
        ],
        "title": "MarketSignalOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "https://REPLACE_WITH_PUBLIC_API_HOST",
      "description": "Production HTTPS (set after R3 deploy)"
    }
  ]
}